fix(jaas) do not redirect to plan limit page on auth errors (#9746)
